packagers: yum: support the async URLGrabber
yum uses python-urlgrabber to grab files, but does not allow setting the
meter for parallel fetch.

The detection if parallel fetch is to be used is not configurable, so we
hack a private member to force blocking download.

The progress callback broke  because of python-urlgrabber or yum change,
hopping this modification will be sufficient to avoid future behaviour
breakage.

Change-Id: I41557b1ed5f44d685e5305909b3914e2fcf4ce0a
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>