services: rhel: support upstart
vdsm developers moved libvirtd to upstart, it is the only service that is
using upstart.

The reason for this move is to provide a watch dog for restarting libvirtd
service when it dies.

Although it could have been implemented using various simpler methods, the
upstart approach was selected.

Implementation of rhel services should prefer upstart services over the
sysvinit services.

Change-Id: If7289ee08f8a397665e2becfadd0a32b3af167d9
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>