packager: yum: refresh yum instance at every transaction
yum has long history, especially at the suck where it holds the
include/exclude data.

as we need to handle runtime change in versionlock, we need to force yum
to read the information again.

I would have expected this information will be re-read at every new
transaction... but apparently not.

Change-Id: I35a39253aab15b2be27e73be6ac89fe14c9d3f2a
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>