minidnf: drop fill_sack on clean
In clean method we called fill_sack right after
expiring metadata. This caused a SIGABRT to be
raised by one of the C libraries used by DNF.

We've not completely understood why this happen but
we can safely drop the call to fill_sack here.
The clear operation is called within a minidnf.transaction
so the sack will be destroyed right after the cleaning.
Next operations are still within another minidnf.transaction so
the sack will be filled again in a fresh new base object.

Change-Id: I05755de3406a811434236e020061bdf3ad1680ff
Bug-Url: https://bugzilla.redhat.com/1361888
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>