plugins: hostname.py: Do not install/update iproute
This plugin is not really mandatory for anything, it's mainly useful for
debugging. Do not do here an intrusive action (installing or updating a
package) that is not really needed. If we do find it installed, which is
quite common, keep existing functionality.

Change-Id: I45ca3152342e5e93a837cae033ee27196f0dc949
Bug-Url: https://bugzilla.redhat.com/1405838
Signed-off-by: Yedidyah Bar David <didi@redhat.com>
(cherry picked from commit 0166f14bf830ac18ad4675179a56d13edbfeff27)