core: Read config files early
It used to be at STAGE_INIT, PRIORITY_HIGH - 10.
A recent patch removed the priority.
So any plugin that happened to setdefault some env key to None,
before otopi ran core/config.py, succeeded. So answer files setting the
default (which is the default for our automatically-generated answer
files, so most automation tools inherited this as well), did not work.

Revert this for now.

Later on, we should move to STAGE_BOOT, but change all plugins that
set CONFIG_FILE_NAME to run before us (CORE_CONFIG_INIT).

Change-Id: Ib3adb0821c9d2702092e1bbd80234309888f2838
Signed-off-by: Yedidyah Bar David <didi@redhat.com>