core: Do not write after output is closed
A recent change to add pre/post messages about each event, causes
the machine dialog plugin to try and output such a post event after the
output is closed. This causes otopi to fail.

In the host-deploy context, this is hard to debug, because the log
is sent prior to this stage, thus does not contain the error.

Change-Id: I9b3867014ee7fb70a27fec5d4e6fdebab128bed7
Signed-off-by: Yedidyah Bar David <didi@redhat.com>