dialog: human: Always add answers to generated answer files
E.g. even if they are a result of accepting a default.

Not sure we want this patch at all:

- One might claim that if you run with accept-defaults, and then use
the generated answer file without accept-defaults, you should be asked,
and not to automatically accept the defaults of the run that generated
the answer file.

- OTOH it can be useful as a quick way to generate an answer file to be
used for further editing and refinement.

- It's also useful for automated testing - to run unattended e.g. with
accept-defaults and see what file is generated. This is the reason for
sending this patch currently. However, it still does not include answers
to questions not asked at all, usually because answers for them were
provided in an answer file.

- Perhaps it should be made optional. We need to get some experience
with this feature before deciding.

This patch is for the human dialect only. The machine dialect already
behaves like this. If we decide to make this optional, we probably need
to patch that one as well.

Change-Id: I846004e394c40cb06c895eba7c481b687f6b8f8a
Signed-off-by: Yedidyah Bar David <didi@redhat.com>