filetransaction: Fix str/bytes issues on python 3
FileTransaction tried to handle handle both text (str) and binary data
(bytes) when comparing, writing, and copying file, failing on python 3.
There is no need for this complexity since files can always be handled
as binary data.

Change to always convert the content to binary data. This make
comparing, writing and copying files easier and compatible with python
3.

With this change we can add a host running Fedora 28.

Change-Id: Ib587505ab06f662683730d502246c8b5315ebf2f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>