build: Always match maven version to package version
When I last bumped master version, from 1.7 to 1.8, I forgot to also
change maven version, and it was merged as 1.8.6. But no 1.8 was ever
built or released in Maven Central, so it should be safe to "downgrade"
the version to 1.8.0.

Preventing such errors in the future by making maven version equal to
package version. This means we no longer intend to have the option to
have different versions in maven than the released ones.

Change-Id: Idaffc1d4d38621665b67624a0b0d383f82b74556
Signed-off-by: Yedidyah Bar David <didi@redhat.com>