core: Disable minidnf by default on el7
This is an addition to a previous patch that disables dnf packager.

It is needed because some otopi users import minidnf directly, some of
them even without running otopi. Therefore, we can't use env to force
enabling, like in previous patch. So, to enforce trying dnf, set in the
shell environment: OTOPI_DNF_ENABLE=1

Move the function ok_to_use_dnf from a private one (in previous patch,
in dnfpackager plugin) to a public one in packager module. We can't put
it in minidnf module itself, because it might fail importing dnf.

Change-Id: I1771ca514def80bcf58331f971f02ea03f15a012
Bug-Url: https://bugzilla.redhat.com/1694423
Signed-off-by: Yedidyah Bar David <didi@redhat.com>