minidnf: fix rollback
- Replace code that used dnf methods that were removed - open_history
and _history_undo_operations - with code based on the current code of
dnf's 'history undo' command.

- Remove the Epoch that was prepended to packages' 'display_name'. Not
sure what it was used for, but in practice was likely never used in
places where we care. Right now, it didn't work - searching for packages
with the epoch prepended does not find them.

- Fix available and reinstall_available calculations, so that
engine-setup can correctly check them and prompt 'Setup will not be able
to rollback new packages' if needed. Without this, if packages were not
available in enabled repos, engine-setup didn't prompt, and on rollback
removed the updated packages but didn't install the old ones.

Bug-Url: https://bugzilla.redhat.com/2003441
Change-Id: Ib3a46250158863e95aa7c0ebcb9ea548076c4a4a
Signed-off-by: Yedidyah Bar David <didi@redhat.com>