dialog: human: Allow confirmation via answer file
The dialog classes have a little-used method called 'confirm'.
In the patches for adding answer-file support, 'confirm' wasn't handled.
Do this now, simply by replacing the internal implementation with a call
to queryString.

The only user of dialog.confirm that I know of is minidnf, prompting to
confirm use of a gpg key if RPMs are signed with an unknown key and a
key is provided with the 'gpgkey' configuration item. A next patch will
use this for automating testing of this dnf/rpm feature.

Change-Id: I6a7522b26b7bb0ecdf94440403cb54f262ac77f9
Signed-off-by: Yedidyah Bar David <didi@redhat.com>