core: Filter out secret questions' answers
Add support for env[LOG_FILTER_QUESTIONS] - questions names' for which
the answers are filtered out from logs.

It's initialized to an empty list.

It's used only in Stages.CORE_LOG_INIT, so questions should be added to
it beforehand. This isn't a real limitation, because otherwise we risk
logging the answers anyway.

It's added for engine-setup which will use it in its own patch.

Change-Id: I3feb99aa920aa552081ea0763545d75feb017c34
Signed-off-by: Yedidyah Bar David <didi@redhat.com>