dialog: human: Wrap notes output nicely
If dialog.note is passed long lines of text, it prepends some whitespace
to each line to make output "nice", but does not try to break it, thus
letting it wrap "normally" which looks ugly.

Instead, break lines according to terminal width and prepend the same
whitespace to each part.

Change-Id: I4706323a95103eff3257507dd1bd0fdbaa59cd74
Signed-off-by: Yedidyah Bar David <didi@redhat.com>